Create CSVMaker
I'm getting errors as indicated in the code probably because not every game has all of the rest of these keys.  I was thinking of doing an If statement to check if the key exists then write it, if not then continue but I wanted to run this by you first.